# api/openapi.yaml
openapi: 3.1.0
info:
  title: Personal Workbench API
  version: 0.2.0
  description: Local loopback API owned by the workbenchd sidecar.
  license:
    name: Private use
    identifier: LicenseRef-Private
servers:
  - url: http://127.0.0.1:{port}
    variables:
      port:
        default: "0"
security:
  - bearerAuth: []
paths:
...
  /api/v2/meta:
    get:
      summary: Get service and workspace metadata
      operationId: getMeta
      responses:
        "200":
          description: Service metadata
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Meta"
        default:
          $ref: "#/components/responses/Problem"
  /api/v2/preferences:
    get:
      summary: Read workspace-scoped UI preferences
      operationId: getPreferences
      responses:
        "200":
          description: Workspace-scoped UI preferences
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Preferences"
        default:
          $ref: "#/components/responses/Problem"
    patch:
      summary: Update workspace-scoped UI preferences
      operationId: updatePreferences
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PreferencesUpdate"
      responses:
        "200":
          description: Updated workspace-scoped UI preferences
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Preferences"
        default:
          $ref: "#/components/responses/Problem"
  /api/v2/dashboard:
    get:
      summary: Get the today dashboard
      operationId: getDashboard
      parameters:
        - $ref: "#/components/parameters/Timezone"
      responses:
        "200":
          description: Today dashboard
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Dashboard"
        default:
          $ref: "#/components/responses/Problem"
  /api/v2/archive-types:
    get:
      summary: List custom archive types and field definitions
      operationId: listArchiveTypes
      responses:
        "200":
          description: Archive type definitions
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ArchiveTypeDefinition"
        default:
          $ref: "#/components/responses/Problem"
    post:
      summary: Create a custom archive type
      operationId: createArchiveType
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/ArchiveTypeInput" }
      responses:
        "201":
          description: Created archive type
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ArchiveTypeDefinition" }
        default: { $ref: "#/components/responses/Problem" }
  /api/v2/archive-types/{typeId}:
    parameters:
      - $ref: "#/components/parameters/ArchiveTypeId"
    get:
      summary: Get a custom archive type
      operationId: getArchiveType
      responses:
        "200":
          description: Archive type
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ArchiveTypeDefinition" }
        default: { $ref: "#/components/responses/Problem" }
    patch:
      summary: Update a custom archive type
      operationId: updateArchiveType
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/ArchiveTypeInput" }
      responses:
        "200":
          description: Updated archive type
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ArchiveTypeDefinition" }
        default: { $ref: "#/components/responses/Problem" }
    delete:
      summary: Delete an unused archive type
      operationId: deleteArchiveType
      responses:
        "204": { description: Archive type deleted }
        default: { $ref: "#/components/responses/Problem" }
  /api/v2/archive-types/{typeId}/fields:
    parameters:
      - $ref: "#/components/parameters/ArchiveTypeId"
    post:
      summary: Add a custom field
      operationId: createArchiveField
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/ArchiveFieldInput" }
      responses:
        "201":
          description: Created custom field
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ArchiveFieldDefinition" }
        default: { $ref: "#/components/responses/Problem" }
  /api/v2/archive-fields/{fieldId}:
    parameters:
      - name: fieldId
        in: path
        required: true
        schema: { type: string }
    patch:
      summary: Update a custom field
      operationId: updateArchiveField
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/ArchiveFieldInput" }
      responses:
        "200":
          description: Updated custom field
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ArchiveFieldDefinition" }
        default: { $ref: "#/components/responses/Problem" }
    delete:
      summary: Delete a custom field and its values
      operationId: deleteArchiveField
      responses:
        "204": { description: Custom field deleted }
        default: { $ref: "#/components/responses/Problem" }
  /api/v2/archives:
    get:
      summary: List and filter archives
      operationId: listArchives
      parameters:
        - $ref: "#/components/parameters/Query"
        - name: typeId
          in: query
          schema: { type: string }
        - name: sort
          in: query
          schema: { type: string, enum: [updated, title], default: updated }
        - $ref: "#/components/parameters/Limit"
        - $ref: "#/components/parameters/Offset"
      responses:
        "200":
          description: Archive page
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ArchivePage"
        default:
          $ref: "#/components/responses/Problem"
    post:
      summary: Create an archive
      operationId: createArchive
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ArchiveInput"
      responses:
        "201":
          description: Created archive
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Archive"
        default:
          $ref: "#/components/responses/Problem"
  /api/v2/archives/{archiveId}:
    parameters:
      - $ref: "#/components/parameters/ArchiveId"
    get:
      summary: Get an archive
      operationId: getArchive
      responses:
        "200":
          description: Archive details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Archive"
        default:
          $ref: "#/components/responses/Problem"
    patch:
      summary: Update an archive
      operationId: updateArchive
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ArchiveInput"
      responses:
        "200":
          description: Updated archive
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Archive"
        default:
          $ref: "#/components/responses/Problem"
    delete:
      summary: Move an archive to trash
      operationId: trashArchive
      responses:
        "204":
          description: Moved to trash
        default:
          $ref: "#/components/responses/Problem"
  /api/v2/archives/{archiveId}/relations:
    parameters:
      - $ref: "#/components/parameters/ArchiveId"
    get:
      summary: List archive relations
      operationId: listArchiveRelations
      responses:
        "200":
          description: Archive relations
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/Relation" }
        default: { $ref: "#/components/responses/Problem" }
    post:
      summary: Create an archive relation
      operationId: createArchiveRelation
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/RelationInput" }
      responses:
        "201":
          description: Created relation
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Relation" }
        default: { $ref: "#/components/responses/Problem" }
  /api/v2/relations/{relationId}:
    delete:
      summary: Remove an entity relation
      operationId: deleteRelation
      parameters:
        - name: relationId
          in: path
          required: true
          schema: { type: string, format: uuid }
      responses:
        "204": { description: Relation removed }
        default: { $ref: "#/components/responses/Problem" }
  /api/v2/archives/{archiveId}/attachments:
    parameters:
      - $ref: "#/components/parameters/ArchiveId"
    get:
      summary: List archive attachments
      operationId: listArchiveAttachments
      responses:
        "200":
          description: Managed attachments
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/Attachment" }
        default: { $ref: "#/components/responses/Problem" }
    post:
      summary: Import managed attachments
      operationId: importArchiveAttachments
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/AttachmentImportInput" }
      responses:
        "202":
          description: Attachment import job accepted
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Job" }
        default: { $ref: "#/components/responses/Problem" }
  /api/v2/archives/{archiveId}/activity:
    parameters:
      - $ref: "#/components/parameters/ArchiveId"
    get:
      summary: List archive activity
      operationId: listArchiveActivity
      responses:
        "200":
          description: Archive activity
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/Activity" }
        default: { $ref: "#/components/responses/Problem" }
  /api/v2/attachments/{attachmentId}:
    delete:
      summary: Remove a managed attachment
      operationId: deleteAttachment
      parameters:
        - name: attachmentId
          in: path
          required: true
          schema: { type: string, format: uuid }
      responses:
        "204": { description: Attachment removed }
        default: { $ref: "#/components/responses/Problem" }
  /api/v2/attachments/{attachmentId}/open-target:
    get:
      summary: Resolve a managed attachment for native opening
      operationId: getAttachmentOpenTarget
      parameters:
        - name: attachmentId
          in: path
          required: true
          schema: { type: string, format: uuid }
      responses:
        "200":
          description: Validated local target
          content:
            application/json:
              schema:
                type: object
                required: [path]
                properties:
                  path: { type: string }
        default: { $ref: "#/components/responses/Problem" }
  /api/v2/tasks:
    get:
      summary: List unified tasks for a view, calendar range, or archive
      operationId: listTasks
      parameters:
        - name: view
          in: query
          schema:
            type: string
            enum: [today, tomorrow, all, completed]
            default: all
        - $ref: "#/components/parameters/Timezone"
        - $ref: "#/components/parameters/Query"
        - name: archiveId
          in: query
          schema: { type: string, format: uuid }
        - name: from
          in: query
          schema: { type: string, format: date-time }
        - name: to
          in: query
          schema: { type: string, format: date-time }
      responses:
        "200":
          description: Tasks
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Task"
        default:
          $ref: "#/components/responses/Problem"
    post:
      summary: Create a task
      operationId: createTask
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TaskInput"
      responses:
        "201":
          description: Created task
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Task"
        default:
          $ref: "#/components/responses/Problem"
  /api/v2/tasks/{taskId}:
    parameters:
      - name: taskId
        in: path
        required: true
        schema:
          type: string
          format: uuid
    get:
      summary: Get task details for the inspector
      operationId: getTask
      responses:
        "200":
          description: Task details
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Task" }
        default: { $ref: "#/components/responses/Problem" }
    patch:
      summary: Update a task
      operationId: updateTask
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TaskInput"
      responses:
        "200":
          description: Updated task
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Task"
        default:
          $ref: "#/components/responses/Problem"
    delete:
      summary: Move a task to trash
      operationId: trashTask
      responses:
        "204":
          description: Moved to trash
        default:
          $ref: "#/components/responses/Problem"
  /api/v2/search:
    get:
      summary: Search workspace entities
      operationId: search
      parameters:
        - $ref: "#/components/parameters/QueryRequired"
      responses:
        "200":
          description: Search results
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/SearchResult"
        default:
          $ref: "#/components/responses/Problem"
...
  /internal/shutdown:
    post:
      summary: Request graceful sidecar shutdown
      operationId: shutdown
      responses:
        "202":
          description: Shutdown accepted
        default:
          $ref: "#/components/responses/Problem"
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
...
  responses:
    Problem:
      description: RFC 9457 compatible problem detail
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/Problem"
  schemas:
...
    Meta:
      type: object
      required: [apiVersion, serviceVersion, workspaceName, schemaVersion]
      properties:
        apiVersion: { type: integer }
        serviceVersion: { type: string }
        workspaceName: { type: string }
        schemaVersion: { type: integer }
...
    Preferences:
      type: object
      required: [theme, sidebarCollapsed, inspectorWidth, recentSearches]
      properties:
        theme: { type: string, enum: [light, dark, system] }
        sidebarCollapsed: { type: boolean }
        inspectorWidth: { type: integer, minimum: 300, maximum: 480 }
        recentSearches:
          type: array
          maxItems: 8
          items: { $ref: "#/components/schemas/RecentSearch" }
    PreferencesUpdate:
      type: object
      additionalProperties: false
      properties:
        theme: { type: string, enum: [light, dark, system] }
        sidebarCollapsed: { type: boolean }
        inspectorWidth: { type: integer, minimum: 300, maximum: 480 }
        recentSearches:
          type: array
          maxItems: 8
          items: { $ref: "#/components/schemas/RecentSearch" }
...